@@ -3,11 +3,10 @@ library widget_circular_animator;
 import 'dart:math';
 
 import 'package:flutter/material.dart';
-import 'package:meta/meta.dart';
 
 class WidgetCircularAnimator extends StatefulWidget {
   const WidgetCircularAnimator({
-    @required this.child,
+    required this.child,
     this.innerColor = Colors.deepOrange,
     this.outerColor = Colors.deepOrange,
     this.innerAnimation = Curves.linear,
@@ -18,7 +17,7 @@ class WidgetCircularAnimator extends StatefulWidget {
     this.innerAnimationSeconds = 30,
     this.outerAnimationSeconds = 30,
     this.reverse = true,
-  }) : assert(child != null);
+  });
 
   final Color innerColor;
   final Color outerColor;
@@ -38,10 +37,10 @@ class WidgetCircularAnimator extends StatefulWidget {
 
 class _WidgetAnimatorState extends State<WidgetCircularAnimator>
     with TickerProviderStateMixin {
-  Animation<double> animation1;
-  Animation<double> animation2;
-  AnimationController controller2;
-  AnimationController controller1;
+  late Animation<double> animation1;
+  late Animation<double> animation2;
+  late AnimationController controller2;
+  late AnimationController controller1;
 
   @override
   void initState() {
@@ -141,13 +140,13 @@ class _WidgetAnimatorState extends State<WidgetCircularAnimator>
 class Arc2Painter extends CustomPainter {
   Arc2Painter({this.color, this.iconsSize = 3});
 
-  final Color color;
+  final Color? color;
   final double iconsSize;
 
   @override
   void paint(Canvas canvas, Size size) {
     final Paint p = Paint()
-      ..color = color
+      ..color = color!
       ..strokeWidth = 1.0
       ..strokeCap = StrokeCap.round
       ..style = PaintingStyle.stroke;
@@ -174,6 +173,7 @@ class Arc2Painter extends CustomPainter {
         Offset(centerX + lineLength, centerY - lineLength), p);
     canvas.drawLine(Offset(centerX + lineLength, centerY + lineLength),
         Offset(centerX - lineLength, centerY - lineLength), p);
+
     // the circle
     canvas.drawCircle(Offset(centerX, centerY), iconsSize + 1, p);
 
@@ -193,13 +193,13 @@ class Arc2Painter extends CustomPainter {
 class Arc1Painter extends CustomPainter {
   Arc1Painter({this.color, this.iconsSize = 3});
 
-  final Color color;
+  final Color? color;
   final double iconsSize;
 
   @override
   void paint(Canvas canvas, Size size) {
     final Paint p = Paint()
-      ..color = color
+      ..color = color!
       ..strokeWidth = 1
       ..strokeCap = StrokeCap.round
       ..style = PaintingStyle.stroke;
