@@ -1,3 +1,4 @@
+import 'package:example/Res.dart';
 import 'package:flutter/material.dart';
 import 'package:widget_circular_animator/widget_circular_animator.dart';
 
@@ -10,29 +11,26 @@ class MyApp extends StatelessWidget {
   Widget build(BuildContext context) {
     return MaterialApp(
       home: Scaffold(
-        body: Center(
-            child: WidgetCircularAnimator(
-          size: 200,
+          body: Center(
+        child: WidgetCircularAnimator(
+          size: 250,
           innerIconsSize: 3,
           outerIconsSize: 3,
-          innerAnimation: Curves.bounceIn,
-          outerAnimation: Curves.bounceIn,
-          innerColor: Colors.orangeAccent,
-          reverse: false,
+          innerAnimation: Curves.easeInOutBack,
+          outerAnimation: Curves.easeInOutBack,
+          innerColor: Colors.deepPurple,
           outerColor: Colors.orangeAccent,
           innerAnimationSeconds: 10,
           outerAnimationSeconds: 10,
           child: Container(
             decoration:
                 BoxDecoration(shape: BoxShape.circle, color: Colors.grey[200]),
-            child: Icon(
-              Icons.person_outline,
-              color: Colors.deepOrange[200],
-              size: 60,
+            child: Image.asset(
+              Resources.person,
             ),
           ),
-        )),
-      ),
+        ),
+      )),
     );
   }
 }
